Moved local file logic to a separate class
